Make broadcasting fault tolerant with RPC manager and context
Create a RPC manager actor to funnel IO through, so that we can store handlers to retry.
Also wrap the node actor, manager actor, and a generic state actor (for challenge specific state) into a context object,
so we can have a uniform interface for handlers. I'm a little unhappy with the RPC manager and context code organisation,
but that can be refactored separately.